refactor: UI 로직 App 클래스에서 분리
